requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
neo4j==5.14.0
python-dotenv==1.0.0
pydantic==2.6.1
qdrant-client==1.7.0
sentence-transformers==2.3.1
graphql-core==3.2.3
strawberry-graphql==0.219.2
python-jose==3.3.0
passlib==1.7.4
pytest==8.0.0
httpx==0.26.0
streamlit==1.32.0
networkx==3.2.1
plotly==5.19.0 